Changed expiration field from an integer field to float field to deal with storing infinity. Abstracted getting an existing message from a channel to a function in utils. Added comments for future changes.

+++ rsvp_bot/models.py
@@ -39,5 +39,5 @@ class EventBoard(models.Model):
     channel = models.ForeignKey(Channel, on_delete=models.CASCADE)
     message = models.ForeignKey(Message, on_delete=models.CASCADE)
     # expirations is time in seconds before event expires
-    expiration = models.IntegerField()
+    expiration = models.FloatField()
     events = models.ManyToManyField(Event)
